Add files via upload
In this Project, we'll work with data provided by the United States Social Security Administration, which lists first names along with the number and sex of babies they were given to in each year.
The data spans 101 years, from 1920 to 2020.
The goal is to understand the trends in popularity of first names. Comparing between classic names and more recent names.