Add files via upload
This Project consists of a statistical analysis between men's and women's soccer matches. The Python libraries used were pandas, matplotlib, pingouin and scipy.stats (optional).

The goal of the Project is to run an hypothesis test that allows to answer if more goals are scored in women's soccer matches than men's.